+++ src/lib/types/enums.ts
@@ -1,0 +1,14 @@
+export enum NavLinks{
+     Sounds = 'sounds',
+     BreathingExercise = 'exercise',
+     Pomodoro = 'pomodoro'
+}
+export enum UpdaterStatus {
+     Checking = "checking",
+     Updating = "updating",
+     NeedsUpdate = "needs-update",
+     Updated = "updated",
+     CheckError = "failed-check",
+     Completed = "completed",
+     UpdateError = "failed-update"
+}